+++ example/nextjs-app/tailwind.config.js
@@ -1,7 +1,11 @@
 module.exports = {
-  important: '#__next',
+  important: ':is(#__next, .MuiPopover-root)',
   presets: [require('@4design/for-ui/tailwind.config.base.js')],
-  content: ['./pages/**/*.{js,ts,jsx,tsx}', './components/**/*.{js,ts,jsx,tsx}'],
+  content: [
+    './pages/**/*.{js,ts,jsx,tsx}',
+    './components/**/*.{js,ts,jsx,tsx}',
+    './node_modules/@4design/for-ui/dist/**/*.js',
+  ],
   theme: {
     extend: {},
   },