+++ example/vite-app/src/components/ui-parts/map/index.tsx
@@ -1,3 +1,0 @@
-export * from './GoogleMapsProvider';
-export * from './Marker';
-export * from './MarkerCluster';